+++ src/lib.rs
@@ -5,8 +5,8 @@ pub mod merkle_tree;
 pub mod poseidon;
 pub mod recursion;
 // pub mod rollup;
-pub mod new_transaction;
 pub mod sparse_merkle_tree;
+pub mod transaction;
 pub mod zkdsa;
 
 pub extern crate plonky2;
