+++ src/transaction/gadgets/deposit_info/mod.rs
@@ -7,7 +7,7 @@ use plonky2::{
 use serde::{Deserialize, Serialize};
 
 use crate::{
-    new_transaction::asset::VariableIndex,
+    transaction::asset::VariableIndex,
     zkdsa::{account::Address, gadgets::account::AddressTarget},
 };
